@@ -292,45 +294,52 @@
     "@capacitor/local-notifications": "^7.0.1",
     "@capacitor/splash-screen": "^7.0.1",
     "@capacitor/status-bar": "^7.0.1",
-    "@capawesome-team/capacitor-file-opener": "^5.0.0",
-    "@capawesome/capacitor-file-picker": "^5.0.0",
-    "@hugotomazi/capacitor-navigation-bar": "^2.0.1",
+    "@capawesome-team/capacitor-file-opener": "^7.0.1",
+    "@capawesome/capacitor-file-picker": "^7.0.1",
+    "@hugotomazi/capacitor-navigation-bar": "^4.0.1",
     "@iconify/vue": "^4.3.0",
     "@playwright/test": "1.47.2",
-    "@types/discord-rpc": "^4.0.3",
-    "@types/file-saver": "^2.0.5",
-    "@types/ip": "^1.1.0",
-    "@types/mime-types": "^2.1.1",
-    "@types/node": "^22.15.3",
-    "@types/pidusage": "^2.0.2",
+    "@types/file-saver": "^2.0.7",
+    "@types/ip": "^1.1.3",
+    "@types/mime-types": "^2.1.4",
+    "@types/node": "^22.15.17",
+    "@types/pidusage": "^2.0.5",
+    "@types/semver": "^7.7.0",
     "@types/sharp": "^0.31.1",
     "@types/source-map-support": "^0.5.6",
-    "@types/three": "^0.150.0",
+    "@types/three": "^0.176.0",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^5.59.1",
-    "@typescript-eslint/parser": "^5.59.1",
-    "@vitejs/plugin-vue": "^5.2.3",
-    "@vitest/browser": "^3.1.2",
-    "@vue/eslint-config-typescript": "^11.0.3",
-    "concurrently": "^8.0.1",
+    "@typescript-eslint/eslint-plugin": "^8.32.0",
+    "@typescript-eslint/parser": "^8.32.0",
+    "@typescript-eslint/utils": "^8.32.0",
+    "@vitejs/plugin-vue": "^5.2.4",
+    "@vitest/browser": "^3.1.3",
+    "@vue/eslint-config-typescript": "^14.5.0",
+    "concurrently": "^9.1.2",
     "cross-env": "^7.0.3",
-    "electron": "^24.1.3",
-    "electron-builder": "^26.0.12",
-    "electronmon": "^2.0.2",
-    "eslint": "^8.39.0",
+    "dmg-builder": "^26.0.15",
+    "electron": "^36.2.0",
+    "electron-builder": "^26.0.15",
+    "electron-builder-squirrel-windows": "^26.0.15",
+    "electronmon": "^2.0.3",
+    "eslint": "^9.26.0",
+    "eslint-config-prettier": "^10.1.5",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-import": "^2.31.0",
-    "eslint-plugin-vue": "^9.11.0",
-    "husky": "^8.0.3",
-    "ip": "^1.1.8",
+    "eslint-plugin-vue": "^10.1.0",
+    "globals": "^16.1.0",
+    "husky": "^9.1.7",
+    "ip": "^2.0.1",
+    "jiti": "^2.4.2",
     "playwright": "1.47.2",
     "postcss": "^8.5.3",
     "prettier": "^2.8.8",
-    "rimraf": "^5.0.0",
+    "rimraf": "^6.0.1",
     "tailwindcss": "^3.3.2",
     "ts-node": "^10.9.2",
-    "tsx": "^3.12.6",
-    "typescript": "^5.0.4",
+    "tsx": "^4.19.4",
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.32.0",
     "vitest": "^3.1.2",
     "vitest-browser-vue": "^0.2.0",
     "vue-tsc": "^1.6.1"
@@ -347,43 +356,45 @@
     "@vueform/slider": "^2.1.9",
     "@vueuse/core": "^10.1.0",
     "audify": "^1.9.0",
-    "chalk": "4.1.2",
+    "chalk": "^5.4.1",
     "chokidar": "^4.0.3",
+    "electron-is-dev": "2.0.0",
     "electron-log": "^4.4.6",
-    "electron-store": "^8.1.0",
+    "electron-store": "^10.0.1",
     "electron-updater": "^6.6.2",
     "electron-window-state": "^5.0.3",
     "file-saver": "^2.0.5",
     "iso-3166-1": "^2.1.1",
-    "iso-639-1": "^3.1.0",
-    "mime-types": "^2.1.35",
+    "iso-639-1": "^3.1.5",
+    "mime-types": "^3.0.1",
     "minimal-discord-rpc": "^1.0.3",
-    "mitt": "^3.0.0",
-    "music-metadata": "^7.13.3",
-    "music-metadata-browser": "^2.5.10",
+    "mitt": "^3.0.1",
+    "music-metadata": "^11.2.1",
     "musicbrainz-api": "^0.25.0",
     "node-vibrant": "^4.0.3",
     "node-wav": "^0.0.2",
-    "open": "8.4.2",
-    "pidusage": "^3.0.2",
-    "pretty-ms": "^8.0.0",
+    "open": "^10.1.2",
+    "pidusage": "^4.0.1",
+    "pretty-ms": "^9.2.0",
     "sharp": "^0.31.3",
     "subsonic-api": "^3.1.2",
     "systeminformation": "^5.25.11",
-    "three": "^0.152.1",
-    "uuid": "^9.0.0",
-    "vite": "^4.3.3",
+    "three": "^0.176.0",
+    "uuid": "^11.1.0",
+    "vite": "^6.3.5",
     "vite-plugin-node-polyfills": "^0.23.0",
     "vite-plugin-windicss": "^1.8.10",
-    "vue": "^3.2.47",
-    "vue-i18n": "^9.8.0",
-    "vue-router": "4",
+    "vue": "^3.5.13",
+    "vue-eslint-parser": "^10.1.3",
+    "vue-i18n": "^11.1.3",
+    "vue-router": "^4.5.1",
     "vue-virtual-scroller": "^2.0.0-beta.8",
     "windicss": "^3.5.6"
   },
   "devEngines": {
-    "node": "18.0.0",
+    "node": ">=18",
     "npm": ">=8.x",
     "yarn": ">=1.22.x"
-  }
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
