@@ -1,5 +1,7 @@
 import { ipcMain } from "electron/main";
-const path = require("node:path");
+import path from "node:path";
+import { __filename, __dirname } from "./utility.js";
+
 const iconName = path.join(__dirname, "../../assets/images/audio64x64.png");
 
 ipcMain.on("ondragstart", (event, filePath) => {
