@@ -1,36 +1,41 @@
 import type { IpcRendererEvent } from "electron";
-import { contextBridge, ipcRenderer } from "electron";
+
+// hack: import electron through cjs so it doesn't throw an error.
+const electron = require('electron');
+const { ipcRenderer, dialog } = electron;
+
 import * as fs from "fs/promises";
 import * as os from "os";
 import * as path from "path";
-import {md5} from "./utility";
+import { md5 } from "./utility.js";
 
-contextBridge.exposeInMainWorld("fs", fs);
-contextBridge.exposeInMainWorld("os", os);
-contextBridge.exposeInMainWorld("path", path);
-contextBridge.exposeInMainWorld("md5", md5);
+window.fs = fs;
+window.os = os;
+window.path = path;
+window.md5 = md5;
+window.dialog = dialog;
 
-contextBridge.exposeInMainWorld("electron", {
-	isMac: os.platform() === "darwin", 
-	isWindows: os.platform() === "win32", 
-	isLinux: os.platform() === "linux",
-	startDrag: (fileName: string) => ipcRenderer.send("ondragstart", fileName),
-	ipcRenderer: {
-		invoke(channel: string, args?: string[]) {
-			return ipcRenderer.invoke(channel, args);
-		},
-		send(channel: string, args: unknown[]) {
-			ipcRenderer.send(channel, args);
-		},
-		on(channel: string, func: (...args: unknown[]) => void) {
-			const subscription = (_event: IpcRendererEvent, ...args: unknown[]) =>
-				func(...args);
-			ipcRenderer.on(channel, subscription);
+window.electron = {
+  startDrag: (fileName: string) => ipcRenderer.send("ondragstart", fileName),
+  isMac: os.platform() === "darwin",
+  isWindows: os.platform() === "win32",
+  isLinux: os.platform() === "linux",
+  ipcRenderer: {
+    invoke<T>(channel: string, args?: unknown[]): Promise<T> {
+      return ipcRenderer.invoke(channel, args);
+    },
+    send(channel: string, args: unknown): void {
+      ipcRenderer.send(channel, args);
+    },
+    on<T>(channel: string, func: (...args: T[]) => void): (() => void) | undefined {
+      const subscription = (_event: IpcRendererEvent, ...args: T[]) => func(...args);
+      ipcRenderer.on(channel, subscription);
+      return () => ipcRenderer.removeListener(channel, subscription);
+    },
+    once<T>(channel: string, func: (...args: T[]) => void): void {
+      ipcRenderer.once(channel, (_event: IpcRendererEvent, ...args: T[]) => func(...args));
+    },
+  },
+};
 
-			return () => ipcRenderer.removeListener(channel, subscription);
-		},
-		once(channel: string, func: (...args: unknown[]) => void) {
-			ipcRenderer.once(channel, (_event, ...args) => func(...args));
-		},
-	},
-});
+export {};
