@@ -3,8 +3,8 @@ import { Queue } from "@/logic/queue";
 import { Track } from "@/logic/track";
 import { useLocalStorage } from "@vueuse/core";
 import { ref, watch } from "vue";
-import { AmethystAudioNodeManager } from "./audioManager";
-import { EventEmitter } from "./eventEmitter";
+import { AmethystAudioNodeManager } from "@/logic/audioManager.ts";
+import { EventEmitter } from "@/logic/eventEmitter.ts";
 import { secondsToColinHuman, secondsToHuman } from "@shared/formating";
 import type { Amethyst } from "@/amethyst";
 import "./analytics";
