@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/consistent-type-imports */
+ 
 declare global {
 
 	interface Import {
@@ -13,7 +13,7 @@ declare global {
   interface Window {
     fs: typeof import("fs/promises");
     os: typeof import("os");
-    md5: typeof import("../../main/utility").md5;
+    md5: typeof import("../../main/utility.js").md5;
     path: typeof import("path");
     dialog: Electron.Dialog;
     electron: {
@@ -22,14 +22,11 @@ declare global {
 			isWindows: boolean, 
 			isLinux: boolean,
 			ipcRenderer: {
-				invoke<T>(channel: string, args?: any[]): Promise<T>
-				send(channel: string, args: unknown[]): void
-				on<T>(
-					channel: string,
-					func: (...args: T[]) => void
-				): (() => void) | undefined
-				once<T>(channel: string, func: (...args: T[]) => void): void
-			}
+				invoke<T>(channel: string, args?: unknown[]): Promise<T>;
+				send(channel: string, args: unknown): void;
+				on<T>(channel: string, func: (...args: T[]) => void): (() => void) | undefined;
+				once<T>(channel: string, func: (...args: T[]) => void): void;
+			};
 		}
   }
 }
